Fixed scripts
Added an updated version of the hand tracking sharing code. its untested but is definitely better then before. It is still unclear if it works as expected the server side code makes this a little strange. But we'll see
also updated it to the newst networking version(pre.4)  removed a lot of "waiting" scripts.